refactor(monitor): move highlightSelectedFeature to viewer (from Editing)
